pyproject.toml
[project]
name = "ksebj-chat-bot-server"
version = "0.1.0"
description = "AI Chatbot Server for Collaboration Tool"
authors = [
    {name = "JaeHyunAn",email = "98042706+yyuneu@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11, <4.0"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "openai (>=1.97.1,<2.0.0)",
    "python-docx (>=1.2.0,<2.0.0)",
    "python-pptx (>=1.0.2,<2.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "langchain-openai (>=0.3.28,<0.4.0)",
    "tavily-python (>=0.7.10,<0.8.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "selenium (>=4.34.2,<5.0.0)",
    "webdriver-manager (>=4.0.2,<5.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "google-api-python-client (>=2.177.0,<3.0.0)",
    "langchain-google-community (>=2.0.7,<3.0.0)",
    "arxiv (>=2.2.0,<3.0.0)",
    "pymupdf (>=1.26.3,<2.0.0)",
    "docxtpl",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

# --- 아래 내용을 추가 ---
[tool.pylint.MASTER]
# 런타임에 src 폴더를 맨 앞에 껴 넣어 줍니다
init-hook = "import sys; sys.path.insert(0, 'src')"

[tool.pylint.format]
max-line-length = 200
